requirements.txt
jdcal==1.0
openpyxl==2.2.4
psutil==3.0.1
pycrypto==2.6.1
python-dateutil==2.4.2
python-qt5==0.1.10
requests==2.6.2
six==1.9.0
